+++ stxt/src/schema/type/ENUM.ts
@@ -16,7 +16,7 @@ export const ENUM: Type = {
         const value = n.getValue();
         const allowed = ndef.getValues(); // ReadonlySet<string>
 
-        if (!allowed.has(value)) {
+        if (!ndef.isAllowedValue(value)) {
             throw new ValidationException(n.getLine(),"INVALID_VALUE",`The value '${value}' not allowed. Only: ${Array.from(allowed).join(", ")}`);
         }
     },