+++ stxt/CHANGELOG.md
@@ -9,6 +9,8 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 - Template values are only allowed when the type is `ENUM`.
 - Parse node `description` of templates
 - Display correct line number when parsing templates
+- ENUM with normalized values
+- ENUM with no repeted values
 
 ## [0.2.0]